+++ edsnlp/processing/distributed.py
@@ -1,12 +1,14 @@
 from functools import partial
-from typing import Dict, List, Tuple, Union
+from typing import Any, Callable, Dict, List, Tuple, Union
 
+from decorator import decorator
 from pyspark.sql import DataFrame, SparkSession
 from pyspark.sql import functions as F
 from pyspark.sql import types as T
 from spacy import Language
 
 from edsnlp.pipelines.base import BaseComponent
+from edsnlp.processing.typing import DataFrameModules, DataFrames, get_module
 
 
 def pyspark_type_finder(obj):
@@ -14,26 +16,48 @@ def pyspark_type_finder(obj):
     Returns (when possible) the PySpark type of any python object
     """
     try:
-        infered_type = T._infer_type(obj)
-        print(f"Infered type is {repr(infered_type)}")
-        return infered_type
+        inferred_type = T._infer_type(obj)
+        print(f"Inferred type is {repr(inferred_type)}")
+        return inferred_type
     except TypeError:
         raise TypeError("Cannot infer type for this object.")
 
 
+@decorator
+def module_checker(
+    func: Callable,
+    *args,
+    **kwargs,
+) -> Any:
+
+    args = list(args)
+    note = args.pop(0)
+    module = get_module(note)
+
+    if module == DataFrameModules.PYSPARK:
+        return func(note, *args, **kwargs)
+    elif module == DataFrameModules.KOALAS:
+        import databricks.koalas  # noqa F401
+
+        note_spark = note.to_spark()
+        note_nlp_spark = func(note_spark, *args, **kwargs)
+        return note_nlp_spark.to_koalas()
+
+
+@module_checker
 def pipe(
-    note: DataFrame,
+    note: DataFrames,
     nlp: Language,
     additional_spans: Union[List[str], str] = "discarded",
     extensions: List[Tuple[str, T.DataType]] = [],
 ) -> DataFrame:
     """
-    Function to apply a spaCy pipe to a pyspark DataFrame note
+    Function to apply a spaCy pipe to a pyspark or koalas DataFrame note
 
     Parameters
     ----------
     note : DataFrame
-        A pyspark DataFrame with a `note_id` and `note_text` column
+        A Pyspark or Koalas DataFrame with a `note_id` and `note_text` column
     nlp : Language
         A spaCy pipe
     additional_spans : Union[List[str], str], by default "discarded"