+++ edsnlp/processing/wrapper.py
@@ -1,23 +1,21 @@
 from typing import Any, Dict, List, Union
 
-import pandas as pd
-import pyspark.sql as ps
 from spacy import Language
 
 from .parallel import pipe as parallel_pipe
 from .simple import ExtensionSchema
 from .simple import pipe as simple_pipe
-from .spark import pipe as spark_pipe
+from .typing import DataFrameModules, DataFrames, get_module
 
 
 def pipe(
-    note: Union[pd.DataFrame, ps.DataFrame],
+    note: DataFrames,
     nlp: Language,
-    how: str = "parallel",
+    n_jobs: int = -2,
     additional_spans: Union[List[str], str] = "discarded",
     extensions: ExtensionSchema = [],
     **kwargs: Dict[str, Any],
-) -> Union[pd.DataFrame, ps.DataFrame]:
+) -> DataFrames:
     """
     Function to apply a spaCy pipe to a pandas or pyspark DataFrame
 
@@ -25,17 +23,17 @@ def pipe(
     Parameters
     ----------
     note : DataFrame
-        A pandas DataFrame with a `note_id` and `note_text` column
+        A pandas/pyspark/koalas DataFrame with a `note_id` and `note_text` column
     nlp : Language
         A spaCy pipe
-    how : str, by default "parallel"
-        3 methods are available here:
+    n_jobs : int, by default -2
+        Only used when providing a Pandas DataFrame
 
-        - `how='simple'`: Single process on a pandas DataFrame
-        - `how='parallel'`: Parallelised processes on a pandas DataFrame
-        - `how='spark'`: Distributed processes on a pyspark DataFrame
+        - `n_jobs=1` corresponds to `simple_pipe`
+        - `n_jobs>1` corresponds to `parallel_pipe` with `n_jobs` parallel workers
+        - `n_jobs=-1` corresponds to `parallel_pipe` with maximun number of workers
+        - `n_jobs=-2` corresponds to `parallel_pipe` with maximun number of workers -1
     additional_spans : Union[List[str], str], by default "discarded"
-
         A name (or list of names) of SpanGroup on which to apply the pipe too:
         SpanGroup are available as `doc.spans[spangroup_name]` and can be generated
         by some pipes. For instance, the `date` pipe populates doc.spans['dates']
@@ -48,56 +46,49 @@ def pipe(
 
     Returns
     -------
-    Union[pd.DataFrame, ps.DataFrame]
+    DataFrame
         A DataFrame with one line per extraction
     """
 
-    if (type(note) == ps.DataFrame) and (how != "spark"):
+    module = get_module(note)
+
+    if module == DataFrameModules.PANDAS:
+        if n_jobs == 1:
+
+            return simple_pipe(
+                note=note,
+                nlp=nlp,
+                additional_spans=additional_spans,
+                extensions=extensions,
+                **kwargs,
+            )
+
+        else:
+
+            return parallel_pipe(
+                note=note,
+                nlp=nlp,
+                additional_spans=additional_spans,
+                extensions=extensions,
+                n_jobs=n_jobs,
+                **kwargs,
+            )
+
+    if extensions and type(extensions) != dict:
         raise ValueError(
-            "You are providing a pyspark DataFrame, please use `how='spark'`"
-        )
-    if how == "simple":
-
-        return simple_pipe(
-            note=note,
-            nlp=nlp,
-            additional_spans=additional_spans,
-            extensions=extensions,
-            **kwargs,
+            """
+            When using Spark or Koalas, you should provide extension names
+            along with the extension type (as a dictionnary):
+            `d[extension_name] = extension_type`
+            """  # noqa W291
         )
 
-    if how == "parallel":
+    from .distributed import pipe as distributed_pipe
 
-        return parallel_pipe(
-            note=note,
-            nlp=nlp,
-            additional_spans=additional_spans,
-            extensions=extensions,
-            **kwargs,
-        )
-
-    if how == "spark":
-        if type(note) == pd.DataFrame:
-            raise ValueError(
-                """
-                You are providing a pandas DataFrame with `how='spark'`,
-                which is incompatible.
-                """
-            )
-
-        if extensions and type(extensions) != dict:
-            raise ValueError(
-                """
-                When using Spark, you should provide extension names
-                along with the extension type (as a dictionnary):
-                `d[extension_name] = extension_type`
-                """  # noqa W291
-            )
-
-        return spark_pipe(
-            note=note,
-            nlp=nlp,
-            additional_spans=additional_spans,
-            extensions=extensions,
-            **kwargs,
-        )
+    return distributed_pipe(
+        note=note,
+        nlp=nlp,
+        additional_spans=additional_spans,
+        extensions=extensions,
+        **kwargs,
+    )