+++ changelog.md
@@ -6,6 +6,8 @@
 - Change the alignment procedure, dropping clumsy `numpy` dependency in favour of `bisect`
 - Change the name of `eds.antecedents` to `eds.history`.
   Calling `eds.antecedents` still works, but issues a deprecation warning and support will be removed in a future version.
+- Add a `eds.covid` component, that identifies mentions of COVID
+- Change the demo, to include NER components
 
 ## v0.4.0